Upgrade cursor handling

+++ Ikala-Graphics/src/main/java/com/ikalagaming/graphics/frontend/gui/data/NextWindowData.java
@@ -7,7 +7,7 @@ import com.ikalagaming.graphics.frontend.gui.flags.RefreshFlags;
 import com.ikalagaming.graphics.frontend.gui.flags.WindowFlags;
 import com.ikalagaming.graphics.frontend.gui.util.Rect;
 
-import org.joml.Vector2f;
+import org.joml.Vector2i;
 
 public class NextWindowData {
     /**
@@ -16,14 +16,14 @@ public class NextWindowData {
     public int fieldFlags;
 
     public Condition positionCondition;
-    public final Vector2f positionValue;
-    public final Vector2f positionPivot;
+    public final Vector2i positionValue;
+    public final Vector2i positionPivot;
     public Condition sizeCondition;
-    public final Vector2f sizeValue;
+    public final Vector2i sizeValue;
     public Condition collapsedCondition;
     public boolean collapsedValue;
-    public final Vector2f contentSizeValue;
-    public final Vector2f scrollValue;
+    public final Vector2i contentSizeValue;
+    public final Vector2i scrollValue;
 
     /**
      * @see com.ikalagaming.graphics.frontend.gui.flags.WindowFlags
@@ -37,7 +37,7 @@ public class NextWindowData {
 
     public final Rect sizeConstraintRect;
     public float backgroundAlpha;
-    public final Vector2f menuBarOffsetMinValue;
+    public final Vector2i menuBarOffsetMinValue;
 
     /**
      * @see com.ikalagaming.graphics.frontend.gui.flags.RefreshFlags
@@ -50,19 +50,19 @@ public class NextWindowData {
     public NextWindowData() {
         fieldFlags = NextWindowFlags.NONE;
         positionCondition = Condition.NONE;
-        positionValue = new Vector2f(0, 0);
-        positionPivot = new Vector2f(0, 0);
+        positionValue = new Vector2i(0, 0);
+        positionPivot = new Vector2i(0, 0);
         sizeCondition = Condition.NONE;
-        sizeValue = new Vector2f(0, 0);
+        sizeValue = new Vector2i(0, 0);
         collapsedCondition = Condition.NONE;
         collapsedValue = false;
-        contentSizeValue = new Vector2f(0, 0);
-        scrollValue = new Vector2f(0, 0);
+        contentSizeValue = new Vector2i(0, 0);
+        scrollValue = new Vector2i(0, 0);
         windowFlags = WindowFlags.NONE;
         childFlags = ChildFlags.NONE;
         sizeConstraintRect = new Rect(0, 0, 0, 0);
         backgroundAlpha = 1.0f;
-        menuBarOffsetMinValue = new Vector2f(0, 0);
+        menuBarOffsetMinValue = new Vector2i(0, 0);
         windowRefreshFlags = RefreshFlags.NONE;
         viewport = null;
     }
